refactor: Move array helpers to helpers folder
